current tri rasterization

+++ src/vectors/barycentric.c
@@ -24,18 +24,18 @@ t_vector2	fast_barycentric(t_point *p, t_barycentric b, t_point v)
 
 t_vector2	barycentric_coordinates(t_point *p, t_point v)
 {
-    float       b1;
-    float       b2;
-    float       b3;
-    float       b4;
+    double       b1;
+    double       b2;
+    double       b3;
+    double       b4;
     t_vector2   w;
 
-    b1 = (float)p[2].y - (float)p[0].y;
-    b2 = (float)p[2].x - (float)p[0].x;
-    b3 = (float)p[1].y - (float)p[0].y;
-    b4 = (float)v.y - (float)p[0].y;
+    b1 = (double)p[2].y - (double)p[0].y;
+    b2 = (double)p[2].x - (double)p[0].x;
+    b3 = (double)p[1].y - (double)p[0].y;
+    b4 = (double)v.y - (double)p[0].y;
 
-    w.x = ((float)p[0].x * b1 + b4 * b2 - v.x * b1) / (b3 * b2 - ((float)p[1].x - (float)p[0].x) * b1);
+    w.x = ((double)p[0].x * b1 + b4 * b2 - v.x * b1) / (b3 * b2 - ((double)p[1].x - (double)p[0].x) * b1);
     w.y = (b4 - w.x * b3) / b1;
     return(w);
 }
